Ability to stop print faster remotely, similar to stop via LCD.
Currently when sitting next to printer it is possible
to stop print immediately by choosing "Stop print"
option on LCD.

When monitoring printer remotely and having a need to stop
print immediately there is no such possibility at all.
M603 which stops the print is pushed into planner and executed
way too late to be useful.

So far the only methods to stop the print were 1) to disconnect
from serial and reconnect causing reset, 2) reset via command
;C2560_RES, 3) issue emergency M112 (which is processed early)

This change adds M603 (stop the print) command to be processed
early, too.

Early enough to be valuable.