PFW-1423 Unload Manually should Restart MMU
The Unload Manually error is raised when the MMU is booting up.
There is currently no handling for a Retry if there is no command running.

Solution to the problem is changing the Retry button to a Restart MMU button.

Restarting the MMU via Error screen works well on my printer