G76: Add PINDA_TEMP_COMP ifdef
If PINDA_TEMP_COMP is not defined then users cannot run this code.
Temperature calibration is required to run G76 when a
PINDA thermistor is present. For users like myself on MK3S/+, temperature
compensation is not possible. The
menu is not shown because the SuperPinda does not require temperature compensation.

Change in memory on MK3S+
Flash: -2834 bytes
SRAM: 0 bytes