PFW-880 Store menu edit limits in PROGMEM and add jump value
If the menu limits are configurable at runtime (for example depend on
EEPROM) then that is also supported. Example is Backlight menu

Add jump value such that when the current editing value is 0, then we can set
the current editing value to some "jump value". For example when editing the nozzle
temperature, it if's at zero, then start with 200°C on the LCD.

To disable the jump value, set it to zero.

Change in memory:
Flash: +224 bytes
SRAM: 0 bytes