mmu: Fix crash when print is paused on toolchange
The solution is to unblock the planner after the print has been saved
into RAM and the command queue has been cleared. This will ensure
the watchdog does not time out (which causes the FW crash).

Additionally this fixes an issue where the printer does not park the
extruder when the print is paused.

Fixes #3989, #3990