Leave user's MMU-loaded filament loaded after first layer cal
Presently, if user runs the first layer calibration with the filament
that's been loaded into the nozzle by MMU, this filament will be
unloaded after the calibration is complete.

This behavior seems unexpected from user experience point of view: if
you load a filament into the nozzle and proceed to first layer
calibration, you did so for a reason and probably don't want it unloaded
after the first try (just like you wouldn't want it unloaded if you did
so without the MMU).

Any other filament that's not been previously loaded into the nozzle is
fair game to unload (as it was before).