Various MMU LCD fixes
Move the Full Screen message from MMU code
to the LCD code. Also don't allow the MMU
code to control LCD updates when running operations.
This should be taken care of by the LCD menu code.

* Fixes an issue where the ramming sequence
is run twice when executing "Load to Nozzle" on the
same slot which is already loaded
* Fixes issue where Load to Nozzle operation will
enable LCD updates, this allows some of the M600
menus to time out to the status screen. That's not supposed to happen
* Fix edge case where the Load to Nozzle operation can
be run with the target temperature set to 0°C

Change in memory:
Flash: -66 bytes
SRAM: 0 bytes