More elegant printer_smodel_check fix, possible fix for older MMU machines
Found a more elegant way to fix the model check.

May have fixed bug for older MMU machines.
Only comparing up to the length of the value from the g-code, would return equal on older MMU machines trying to run g-code sliced without the MMU.
Unfortunately if that is a feature, it will cause the different printer warning.