Delay Z-Axis on Autoload
Prevent z-axis movement when the operators hand is close to the printhead.

Case 1: Temp is already >= target temp
Instead of moving z-axis immediately when the sensor is triggered,
the movement is executed after the fast load sequence. (Just before filament extrusion is expected)

Case 2: Temp is < target temp and preheat is in progress
Instead of moving z-axis immediately when the sensor is triggered,
the movement is executed after the operator confirms the "Press the knob to load filament" message.