Improve manage inactivity
If the steppers are not allowed to be disabled
then manage_inactivity should reset the inactivity timeout
to prevent the steppers from being disabled.

This is how it's done in Marlin 2.1 (but not in Marlin 1)

This way we don't need to refresh previous_millis_cmd manually in a lot of
places. Just set no_stepper_sleep to true, and manage_inactivity
takes care of it.

Change in memory:
Flash: +86 bytes
SRAM: +1 byte